Add 'senses' attribute to syntactic behaviour
Why?
----
In Princeton WordNet not all syntactic behaviours apply to all synsets a
word participates, there is not easy way to fix this in PWN

What?
-----
I propose adding a new attribute 'senses' that links syntactic
behaviours to senses. This is implemented as follows

LMF: A new attribute 'senses' of type 'IDREFS' is added this is optional
(so all existing valid XML LMF files are still valid
JSON: A property 'senses' is added
RDF: A property 'wn:senses' is addred with domain
synsem:SyntacticFrame and range ontolex:LexicalSense